Rework parser tests
